requirements.txt
discord.py
python-dotenv
openai
PyYAML==6.0
dacite==1.6.*
requests
pymongo
PyNaCl